Host now ACKs the host stop command
